+++ service-search/src/main/java/fi/mml/portti/service/search/SearchServiceImpl.java
@@ -6,7 +6,6 @@ import fi.nls.oskari.search.channel.SearchChannel;
 import fi.nls.oskari.search.channel.SearchableChannel;
 import fi.nls.oskari.service.OskariComponentManager;
 import fi.nls.oskari.util.PropertyUtil;
-import org.apache.commons.lang.StringUtils;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
@@ -18,7 +17,7 @@ import java.util.TreeMap;
 public class SearchServiceImpl implements SearchService {
 
     /** logger */
-    private static Logger log = LogFactory.getLogger(SearchServiceImpl.class);
+    private static final Logger LOG = LogFactory.getLogger(SearchServiceImpl.class);
 
     /** Available channels */
     private volatile Map<String, SearchableChannel> availableChannels = null;
@@ -28,28 +27,28 @@ public class SearchServiceImpl implements SearchService {
      */
     protected void initChannels() {
         final TreeMap<String, SearchableChannel> newChannels = new TreeMap<String, SearchableChannel>();
-        log.debug("Initializing search channels");
+        LOG.debug("Initializing search channels");
         final Map<String, SearchChannel> annotatedChannels = OskariComponentManager.getComponentsOfType(SearchChannel.class);
         // get comma separated active channel IDs
         String[] activeChannelIDs = PropertyUtil.getCommaSeparatedList("search.channels");
 
         if (activeChannelIDs.length == 0) {
-            log.warn("No search channels selected. Using all annotated channels");
+            LOG.warn("No search channels selected. Using all annotated channels");
             activeChannelIDs = annotatedChannels.keySet().toArray(new String[0]);
         }
-        log.info("Instantiating search channels:", activeChannelIDs);
+        LOG.info("Instantiating search channels:", activeChannelIDs);
 
         for (String channelID : activeChannelIDs) {
             String cid = channelID.trim();
             SearchableChannel channel = annotatedChannels.get(cid);
             if (channel == null) {
-                log.warn("Couldn't find annotated search channel for ID:", cid,
+                LOG.warn("Couldn't find annotated search channel for ID:", cid,
                         "- Change the searchchannel to extend fi.nls.oskari.search.channelSearchChannel with",
                         "@Oskari(\"[channel id]\") annotation");
                 channel = getLegacyChannel(cid);
             }
             if (channel == null) {
-                log.warn("Couldn't create search channel for ID:", cid);
+                LOG.warn("Couldn't create search channel for ID:", cid);
                 continue;
             }
             newChannels.put(channel.getId(), channel);
@@ -66,7 +65,7 @@ public class SearchServiceImpl implements SearchService {
 
         String className = PropertyUtil.getOptional("search.channel." + cid + ".className");
         if (className == null || className.trim().length() < 1) {
-            log.error("Class name not found for search channel " + cid);
+            LOG.error("Class name not found for search channel " + cid);
             return null;
         }
         try {
@@ -80,79 +79,78 @@ public class SearchServiceImpl implements SearchService {
                     try {
                         // instantiate and register channel
                         SearchableChannel channel = (SearchableChannel)con.newInstance();
-                        log.info("Instantiated search channel class " + className);
+                        channel.init();
+                        LOG.info("Instantiated search channel class " + className);
                         if (!cid.equals(channel.getId())) {
                             // This doesn't actually cause any harm ATM, but it might later on.
-                            log.warn("Non-matching ID for search channel between properties and class: " + cid + " / " + channel.getId());
-                        }
-                        // TODO insert other properties with reflection
-                        // get all properties for channel excluding className
-                        String regex = "^search\\.channel\\." + cid + "\\.(?!className).*\\b";
-                        List<String> propertyKeys = PropertyUtil.getMatchingPropertyNames(regex);
-                        for (String propertyKey : propertyKeys) {
-                            channel.setProperty(propertyKey.substring(StringUtils.ordinalIndexOf(propertyKey, ".", 3)+1), PropertyUtil.get(propertyKey));
+                            LOG.warn("Non-matching ID for search channel between properties and class: " + cid + " / " + channel.getId());
                         }
                         return channel;
                     } catch (InstantiationException ie) {
-                        log.error("Couldn't instantiate class " + className + " for channel (InstantationException)");
+                        LOG.error("Couldn't instantiate class " + className + " for channel (InstantationException)");
                     } catch (IllegalAccessException iae) {
-                        log.error("Couldn't instantiate class " + className + " for channel (IllegalAccessException)");
+                        LOG.error("Couldn't instantiate class " + className + " for channel (IllegalAccessException)");
                     } catch (InvocationTargetException ite) {
-                        log.error("Couldn't instantiate class " + className + " for channel (InvocationTargetException)");
+                        LOG.error("Couldn't instantiate class " + className + " for channel (InvocationTargetException)");
                     }
                 }
-                log.error("Couldn't find a no-args constructor for search channel class " + className);
+                LOG.error("Couldn't find a no-args constructor for search channel class " + className);
             }
         } catch (Exception cnfe) {
-            log.error("Error constructing (legacy) searchchannel: " + cid + " = " + className);
+            LOG.error("Error constructing (legacy) searchchannel: " + cid + " = " + className);
         }
         return null;
     }
 
     public Query doSearch(final SearchCriteria searchCriteria) {
 
-        String searchString = searchCriteria.getSearchString();
-        log.debug("Search string is", searchString);
-        searchCriteria.setSearchString(searchString);
-
         if (availableChannels == null) {
             initChannels();
+        }
+
+        if(searchCriteria.isReverseGeocode()) {
+            LOG.debug("Reverse geocode for (lat:", searchCriteria.getLat(), ", lon:", searchCriteria.getLon(), ")");
         } else {
-            log.debug("Search channels already initialized");
+            LOG.debug("Search string is", searchCriteria.getSearchString());
         }
 
         long fullQueryStartTime = System.currentTimeMillis();
 
-        if(log.isDebugEnabled())
+        if(LOG.isDebugEnabled()) {
             printsc(searchCriteria);
-
-        if(log.isDebugEnabled())
             printAvailableChannels();
+        }
 
-        Query query = new Query();
+        final Query query = new Query();
         query.setSearchCriteria(searchCriteria);
 
-        for (String channel : searchCriteria.getChannels()) {
-            if (availableChannels.containsKey(channel)) {
-                long timeStart = System.currentTimeMillis();
-                SearchableChannel channelImplementation = availableChannels.get(channel);
-                log.debug("Channel", channelImplementation);
-
-                ChannelSearchResult result = handleChannelSearch(
-                        searchCriteria, channelImplementation);
-                log.debug("Result", result);
-                result.setChannelId(channelImplementation.getId());
-
-                query.addChannelSearchResult(result);
-                long timeEnd = System.currentTimeMillis();
-                log.debug("Search query to", channelImplementation.getId(),
-                        "took", (timeEnd - timeStart), "ms",
-                        "- got", result.getNumberOfResults(), "results" );
+        for (String channelId : searchCriteria.getChannels()) {
+            if (!availableChannels.containsKey(channelId)) {
+                continue;
             }
+            long timeStart = System.currentTimeMillis();
+            SearchableChannel channel = availableChannels.get(channelId);
+            if(!channel.isValidSearchTerm(searchCriteria)) {
+                // Skipping
+                LOG.debug("Skipping ", channel.getId(), "- criteria not valid");
+                continue;
+            }
+            ChannelSearchResult result = handleChannelSearch(searchCriteria, channel);
+            int numResults = 0;
+            if(result != null) {
+                LOG.debug("Result", result);
+                result.setChannelId(channel.getId());
+                query.addChannelSearchResult(result);
+                numResults = result.getNumberOfResults();
+            }
+            long timeEnd = System.currentTimeMillis();
+            LOG.debug("Search query to", channel.getId(),
+                    "took", (timeEnd - timeStart), "ms",
+                    "- got", numResults, "results");
         }
 
         long fullQueryEndTime = System.currentTimeMillis();
-        log.debug("Search full query took", (fullQueryEndTime - fullQueryStartTime), "ms" );
+        LOG.debug("Search full query took", (fullQueryEndTime - fullQueryStartTime), "ms");
 
         return query;
     }
@@ -160,29 +158,30 @@ public class SearchServiceImpl implements SearchService {
     /**
      * Handles actual channel search and catches exceptions
      *
-     * @param searchCriteria
-     * @param actualChannel
+     * @param sc
+     * @param channel
      * @return
      */
     private ChannelSearchResult handleChannelSearch(
-            SearchCriteria searchCriteria, SearchableChannel actualChannel)
+            SearchCriteria sc, SearchableChannel channel)
     {
         try {
-            ChannelSearchResult result = actualChannel.doSearch(searchCriteria);
+            ChannelSearchResult result;
+            if(sc.isReverseGeocode()) {
+                result = channel.doSearch(sc.getLat(), sc.getLon(), sc.getSRS());
+            } else {
+                result = channel.doSearch(sc);
+            }
             List<SearchResultItem> items = result.getSearchResultItems();
             // calculate zoom scales etc common fields if we have an annotated (non-legacy) channel
-            if(actualChannel instanceof SearchChannel) {
-                SearchChannel channel = (SearchChannel) actualChannel;
-                for(SearchResultItem item : items) {
-                    channel.calculateCommonFields(item);
-                }
+            for(SearchResultItem item : items) {
+                channel.calculateCommonFields(item);
             }
             return result;
         } catch (Exception e) {
-            log.error(e, "Search query to", actualChannel.getId(),
-                    "failed! Searchstring was '", searchCriteria.getSearchString(), "'" );
+            LOG.error(e, "Search query to", channel.getId(), "failed! Searchstring was '", sc.getSearchString(), "'");
             ChannelSearchResult result = new ChannelSearchResult();
-            result.setChannelId(actualChannel.getId());
+            result.setChannelId(channel.getId());
             result.setQueryFailed(true);
             return result;
         }
@@ -209,53 +208,53 @@ public class SearchServiceImpl implements SearchService {
     }
 
     private void printsc(SearchCriteria searchCriteria) {
-        log.debug("printing SearchCriteria");
+        LOG.debug("printing SearchCriteria");
 
         try {
-            log.debug("SearchString: " + searchCriteria.getSearchString());
+            LOG.debug("SearchString: " + searchCriteria.getSearchString());
 
             if (searchCriteria.getFromDate() == null) {
-                log.debug("from date = null");
+                LOG.debug("from date = null");
             } else {
-                log.debug("Datefrom: " + searchCriteria.getFromDate().toString());
+                LOG.debug("Datefrom: " + searchCriteria.getFromDate().toString());
             }
 
             if (searchCriteria.getToDate() == null) {
-                log.debug("from to = null");
+                LOG.debug("from to = null");
             } else {
-                log.debug("Dateto: " + searchCriteria.getToDate().toString());
+                LOG.debug("Dateto: " + searchCriteria.getToDate().toString());
             }
 
             for (String cha : searchCriteria.getChannels()) {
-                log.debug("channel for searching: " + cha);
+                LOG.debug("channel for searching: " + cha);
             }
 
-            log.debug("printing parameters");
+            LOG.debug("printing parameters");
             java.util.Collection<String> set = searchCriteria.getParams().keySet();
 
             for (java.util.Iterator<String> iterator = set.iterator(); iterator.hasNext(); ) {
-                log.debug("parm key: " + (String) iterator.next());
+                LOG.debug("parm key: " + (String) iterator.next());
 
             }
         } catch (Exception e) {
-            log.debug("sc error");
+            LOG.debug("sc error");
             e.printStackTrace();
         }
 
-        log.debug("/printing SearchCriteria");
+        LOG.debug("/printing SearchCriteria");
     }
 
     private void printAvailableChannels() {
-        log.debug("printing AvailableChannels");
+        LOG.debug("printing AvailableChannels");
         try {
             java.util.Collection<String> set = availableChannels.keySet();
             for (java.util.Iterator<String> iterator = set.iterator(); iterator.hasNext(); ) {
-                log.debug("channel key: " + (String) iterator.next());
+                LOG.debug("channel key: " + (String) iterator.next());
 
             }
         } catch (Exception e) {
-            log.debug("a error");
+            LOG.debug("a error");
         }
-        log.debug("/printing AvailableChannels");
+        LOG.debug("/printing AvailableChannels");
     }
 }