+++ service-search/src/main/java/fi/nls/oskari/search/channel/SearchableChannel.java
@@ -3,6 +3,7 @@ package fi.nls.oskari.search.channel;
 import fi.mml.portti.service.search.ChannelSearchResult;
 import fi.mml.portti.service.search.IllegalSearchCriteriaException;
 import fi.mml.portti.service.search.SearchCriteria;
+import fi.mml.portti.service.search.SearchResultItem;
 
 /**
  * Interface to search service of a single channel.
@@ -14,19 +15,45 @@ public interface SearchableChannel {
      * @return
      * @throws IllegalSearchCriteriaException
      */
-	public ChannelSearchResult doSearch(SearchCriteria searchCriteria) throws IllegalSearchCriteriaException;
+	ChannelSearchResult doSearch(SearchCriteria searchCriteria) throws IllegalSearchCriteriaException;
 
     /**
      * Returns an id for the search channel
      * @return
      */
-    public String getId();
+    String getId();
 
     /**
-     * Hook method to setup properties for the
-     * @param propertyName
-     * @param propertyValue
+     * Called when channel is initialized
      */
-    @Deprecated
-    public void setProperty(String propertyName, String propertyValue);
+    void init();
+
+    /**
+     * Should return true if channel implements reverse geocoding
+     * @return
+     */
+    boolean hasReverseGeocode();
+
+    /**
+     * Validates the input if its usable for this channel
+     * @param criteria
+     * @return
+     */
+    boolean isValidSearchTerm(SearchCriteria criteria);
+
+    /**
+     * Make reverse geocoding using the channel implementation
+     * @param lon
+     * @param lat
+     * @param srs
+     * @return
+     * @throws IllegalSearchCriteriaException
+     */
+    ChannelSearchResult doSearch(double lon, double lat, final String srs) throws IllegalSearchCriteriaException;
+
+    /**
+     * Setup zoomlevel based on item type etc
+     * @param item
+     */
+    void calculateCommonFields(final SearchResultItem item);
 }